third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/font-size-expected.txt

This is a testharness.js-based test.
Found 45 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Can set 'font-min-size' to CSS-wide keywords: initial
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to CSS-wide keywords: inherit
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to CSS-wide keywords: unset
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to CSS-wide keywords: revert
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to var() references:  var(--A)
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to the 'xx-small' keyword: xx-small
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to the 'x-small' keyword: x-small
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to the 'small' keyword: small
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to the 'medium' keyword: medium
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to the 'large' keyword: large
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to the 'x-large' keyword: x-large
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to the 'xx-large' keyword: xx-large
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to the 'larger' keyword: larger
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to the 'smaller' keyword: smaller
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to a length: 0px
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to a length: -3.14em
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to a length: 3.14cm
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to a length: calc(0px + 0em)
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to a percent: 0%
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to a percent: -3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to a percent: 3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-min-size' to a percent: calc(0% + 0%)
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-min-size
[FAIL] Can set 'font-max-size' to CSS-wide keywords: initial
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to CSS-wide keywords: inherit
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to CSS-wide keywords: unset
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to CSS-wide keywords: revert
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to var() references:  var(--A)
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to the 'infinity' keyword: infinity
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to the 'xx-small' keyword: xx-small
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to the 'x-small' keyword: x-small
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to the 'small' keyword: small
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to the 'medium' keyword: medium
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to the 'large' keyword: large
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to the 'x-large' keyword: x-large
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to the 'xx-large' keyword: xx-large
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to the 'larger' keyword: larger
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to the 'smaller' keyword: smaller
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to a length: 0px
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to a length: -3.14em
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to a length: 3.14cm
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to a length: calc(0px + 0em)
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to a percent: 0%
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to a percent: -3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to a percent: 3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
[FAIL] Can set 'font-max-size' to a percent: calc(0% + 0%)
  Failed to execute 'set' on 'StylePropertyMap': Invalid propertyName: font-max-size
Harness: the test ran to completion.
